+++ cs.js
@@ -953,7 +953,9 @@ export default {
   'pm.form.minimumParticipants': 'Musíte mít alespoň jednoho příjemce.',
   'pm.form.maxParticipants': 'Můžete mít maximálně 10 lidí v konverzaci.',
   // added 2019-02-21
-  'common.title': 'Nadpis'
+  'common.title': 'Nadpis',
+  // added 2019-03-24
+  'workshop.encSettings': 'Nastavení encyklopedie'
 };
 
 /* eslint-enable max-len */